tools/pipette/requirements.txt
# tools/pipette/requirements.txt
pydantic>=2.0
pyyaml>=6.0
